Refactor: Split filter_pet_policies into modules
Refactor the edge function into smaller, more manageable files.
