+++ supabase/functions/filter_pet_policies/cors.ts
@@ -1,0 +1,5 @@
+
+export const corsHeaders = {
+  "Access-Control-Allow-Origin": "*",
+  "Access-Control-Allow-Headers": "authorization, x-client-info, apikey, content-type",
+};
